% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midas.modify.R
\name{Midas.modify}
\alias{Midas.modify}
\title{Modifying MIDAS model}
\usage{
Midas.modify(
  fitted,
  lib.size = NULL,
  mean.rel.abund = NULL,
  mean.rel.abund.1 = NULL,
  taxa.1.prop = NULL
)
}
\arguments{
\item{fitted}{An output from Midas.setup}

\item{lib.size}{A numeric vector of pre-specified library sizes (the length
should be equal to \code{n.sample} if specified).}

\item{mean.rel.abund}{A numeric vector of specified mean relative abundances
for taxa (the length should be equal to \code{n.taxa} in \code{fitted}).}

\item{mean.rel.abund.1}{A numeric vector of specified mean relative abundances
among non-zero samples.}

\item{taxa.1.prop}{A numeric vector of specified proportion of non-zeros for
taxa (the length should be equal to \code{n.taxa} in \code{fitted}). If the
library sizes are given and \code{taxa.1.prop = 'same'}, the proportion of
zero cells in each taxon will remain the same as in the template data.}
}
\value{
Returns a list that updates and adds the following elements to the
input list:
\item{n.sample}{The target sample size in the simulation}
\item{lib.size}{The target library sizes in the simulation}
\item{taxa.1.prop}{The updated taxa non-zero proportions}
\item{sample.1.prop}{The updated proportion of non-zero cells for each subject.}
\item{mu}{The mean values of the multivariate normal distribution in generating
presence-absence data}
\item{eta}{The adjustment to be applied to samples in generating presence-
absence data}
}
\description{
Midas.modify modifies the fitted MIDAS.setup model according to user
specification that one or multiple of the following characteristics, such as
the library sizes, number of samples, proportions of zeros, and the taxa
relative abundances, can be changed. This is useful if the users wants to
introduce an 'effect' in simulation studies.
}
\examples{
library(GUniFrac)
data("throat.otu.tab")
otu.tab = throat.otu.tab[,colSums(throat.otu.tab>0)>1]

fitted = Midas.setup(otu.tab)

# modify library sizes
fitted.modified <- Midas.modify(fitted,
                                lib.size = sample(fitted$lib.size, 2*nrow(otu.tab)) )

# modify proportion of non-zeros in taxa
fitted.modified <- Midas.modify(fitted,
                                taxa.1.prop = sample(fitted$taxa.1.prop))

}
\author{
Mengyu He
}
