% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midas.setup.R
\name{Midas.setup}
\alias{Midas.setup}
\title{Fitting MIDAS model to microbiome data
Midas.setup estimates parameters from a template microbiome dataset (count data
or relative abundance data with zeros) for downstream data simulation.}
\usage{
Midas.setup(otu.tab, n.break.ties = 100, fit.beta = FALSE)
}
\arguments{
\item{otu.tab}{A numeric matrix of template dataset (counts/relative
abundances). Rows are samples, columns are taxa.}

\item{n.break.ties}{The number of replicates used to break ties when ranking
relative abundances. Defaults to \code{100}.}

\item{fit.beta}{A logical indicating whether to fit a beta distribution for
the relative abundance of each taxon. Defaults to \code{FALSE}.}

\item{extra.shrink}{A logical indicating whether to apply extra shrinkage
for the correlation matrix of presence-absence data by estimating \code{alpha}.
Defaults to \code{FALSE}.}
}
\value{
Returns a list that has components:
\item{vec01}{Vectorized presence-absence data}
\item{n.taxa}{The number of taxa in the template data.}
\item{ids}{Ids of taxa presenting in all samples in the template.}
\item{tetra.corr}{The estimated tetrachoric correlation of presence-absence
data of the template.}
\item{corr.rel.corrected}{The estimated Pearson correlation of relative
abundances, which is transformed from Spearman's rank correlation.}
\item{taxa.names}{Names of taxa in the template.}
\item{pbar}{The proportion of non-zeros for each subject.}
\item{prop.1}{The proportion of non-zeros for each taxon.}
\item{obs.rel.abund}{Observed mean relative abundances of each taxon.}
\item{non.zero.rel}{The observed non-zero relative abundances of each taxon.}
\item{n0}{The number of non-zeros for each taxon.}
\item{alpha}{If requested, the estimated \code{alpha} parameter for beta
distribution}
\item{beta}{If requested, the estimated \code{beta} parameter for beta
distribution}
\item{scamfit.lib}{The fitted Shape constrained additive model (SCAM) for
library size}
}
\description{
Fitting MIDAS model to microbiome data
Midas.setup estimates parameters from a template microbiome dataset (count data
or relative abundance data with zeros) for downstream data simulation.
}
\examples{
library(GUniFrac)
data("throat.otu.tab")
otu.tab = throat.otu.tab[,colSums(throat.otu.tab>0)>1]

fitted = Midas.setup(otu.tab)

}
\author{
Mengyu He
}
