% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Midas.sim.R
\name{Midas.sim}
\alias{Midas.sim}
\title{Simulating Realistic Microbiome Data using MIDAS}
\usage{
Midas.sim(fitted.modified, only.rel = FALSE)
}
\arguments{
\item{fitted.modified}{An output from Midas.modify}

\item{only.rel}{A logical indicating whether to only simulate relative-
abundance data. If \code{TRUE}, then the count data will not be generated.
Defaults to \code{FALSE}.}
}
\value{
Returns a list that has components:
\item{sim_01}{A matrix of simulated presence-absence data}
\item{sim_rel}{A matrix of simulated relative-abundance data}
\item{sim_count}{A matrix of simulated count data}
}
\description{
Generate microbiome datasets using parameters from Midas.modify.
}
\examples{
library(GUniFrac)
data("throat.otu.tab")
otu.tab = throat.otu.tab[,colSums(throat.otu.tab>0)>1]

fitted = Midas.setup(otu.tab)
fitted.modified = Midas.modify(fitted)
simu = Midas.sim(fitted.modified, only.rel = FALSE)

}
\author{
Mengyu He
}
