% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIDASim.modify.R
\name{MIDASim.modify}
\alias{MIDASim.modify}
\title{Modifying MIDAS model}
\usage{
MIDASim.modify(
  fitted,
  lib.size = NULL,
  mean.rel.abund = NULL,
  gengamma.mu = NULL
)
}
\arguments{
\item{fitted}{Output from Midas.setup.}

\item{lib.size}{Numeric vector of pre-specified library sizes (length should
be equal to \code{n.sample} if specified).}

\item{mean.rel.abund}{Numeric vector of specified mean relative abundances for
taxa. Length should be equal to \code{n.taxa} in \code{fitted}.}

\item{gengamma.mu}{Numeric vector of specified location parameters for the
parametric model (generalized gamma model). Specify either \code{mean.rel.abund}
or \code{gengamma.mu}, not both. Length should be equal to \code{n.taxa} in \code{fitted}.
See Details.}
}
\value{
Returns an updated list with different elements depending on the value
of \code{fitted$mode}:
\item{n.sample}{Target sample size in the simulation.}
\item{lib.size}{Target library sizes in the simulation.}
\item{taxa.1.prop}{Updated proportions of non-zero values for each taxon.}
\item{sample.1.prop}{Updated proportion of non-zero cells for each subject.}
\item{theta}{Mean values of the multivariate normal distribution in generating
presence-absence data.}
\item{eta}{Adjustment to be applied to samples in generating presence-
absence data.}
\item{mu.est}{Estimated location parameters for the parametric model.}
\item{sigma.est}{Estimated scale parameters for the parametric model.}
\item{Q.est}{Estimated scale parameters for the parametric model.}
}
\description{
Midas.modify modifies the fitted MIDAS.setup model according to user
specification that one or multiple of the following characteristics, such as
the library sizes, taxa relative abundances, location parameters of the parametric
model can be changed. This is useful if the users wants to introduce an 'effect'
in simulation studies.
}
\details{
The parametric model in MIDASim is a location-scale model, specifically, a
generalized gamma model for relative abundances \eqn{\pi} of a taxon. Denote \eqn{t = 1/\pi}.
The generalized gamma distribution for \eqn{t} is chosen so that

\deqn{ln(t)\ =\ \mu\ +\ \sigma \cdot w}

where \eqn{w} follows a log gamma distribution with a shape parameter \eqn{1/Q}.
MIDASim fits the model to the template data and estimates parameters \eqn{\mu}, \eqn{\sigma}
and \eqn{Q} by matching the first two moments of \eqn{\pi} and maximizing the likelihood.
}
\examples{

data("throat.otu.tab")
otu.tab = throat.otu.tab[,colSums(throat.otu.tab>0)>1]

fitted = MIDASim.setup(otu.tab, mode = 'parametric')

# modify library sizes
fitted.modified <- MIDASim.modify(fitted,
                                  lib.size = sample(fitted$lib.size,
                                                    2*nrow(otu.tab)) )

# modify mean relative abundances
fitted.modified <- MIDASim.modify(fitted,
                                  mean.rel.abund = fitted$mean.rel.abund * runif(fitted$n.taxa))
# modify location parameters of the parametric model
fitted.modified <- MIDASim.modify(fitted,
                                  gengamma.mu = fitted$mean.rel.abund * runif(fitted$n.taxa))

}
\author{
Mengyu He
}
